Start evaluator

+++ src/ast/ast.cpp
@@ -27,7 +27,7 @@ std::string Identifier::to_string() const { return value; }
 // }}}
 
 // {{{ IntegerLiteral
-IntegerLiteral::IntegerLiteral(int v) : value(v) {}
+IntegerLiteral::IntegerLiteral(long v) : value(v) {}
 std::string IntegerLiteral::token_literal() const { return "INT"; }
 std::string IntegerLiteral::to_string() const { return std::to_string(value); }
 // }}}
