+++ src/parser/parser_test.cpp
@@ -361,19 +361,19 @@ bool test_boolean_literal_expression() {
 
 bool test_prefix_expression() {
   auto tests = std::vector<std::any>{
-      prefix_test_case<int>{"!5;", token_types::Bang{}, 5},
-      prefix_test_case<int>{"-15;", token_types::Minus{}, 15},
+      prefix_test_case<long>{"!5;", token_types::Bang{}, 5},
+      prefix_test_case<long>{"-15;", token_types::Minus{}, 15},
       prefix_test_case<bool>{"!true;", token_types::Bang{}, true},
       prefix_test_case<bool>{"!false;", token_types::Bang{}, false},
   };
 
   for (size_t i = 0; i < tests.size(); i++) {
     auto test{tests[i]};
-    auto int_test{std::any_cast<prefix_test_case<int>>(&test)};
+    auto int_test{std::any_cast<prefix_test_case<long>>(&test)};
     auto bool_test{std::any_cast<prefix_test_case<bool>>(&test)};
     bool ret_val{true};
     if (int_test) {
-      if (!h_do_prefix_test_case<int, IntegerLiteral>(*int_test)) {
+      if (!h_do_prefix_test_case<long, IntegerLiteral>(*int_test)) {
         ret_val = false;
       }
     } else if (bool_test) {
@@ -395,21 +395,21 @@ bool test_prefix_expression() {
 bool test_infix_expression() {
   using namespace token_types;
   auto tests = std::vector<std::any>{
-      infix_test_case<int>{"3 + 10;", 3, Plus{}, 10},
-      infix_test_case<int>{"5 - 5;", 5, Minus{}, 5},
-      infix_test_case<int>{"5 * 5;", 5, Asterisk{}, 5},
-      infix_test_case<int>{"5 / 5;", 5, Slash{}, 5},
-      infix_test_case<int>{"5 > 5;", 5, GT{}, 5},
-      infix_test_case<int>{"5 < 5;", 5, LT{}, 5},
-      infix_test_case<int>{"5 == 5;", 5, Eq{}, 5},
-      infix_test_case<int>{"5 != 5;", 5, NotEq{}, 5},
+      infix_test_case<long>{"3 + 10;", 3, Plus{}, 10},
+      infix_test_case<long>{"5 - 5;", 5, Minus{}, 5},
+      infix_test_case<long>{"5 * 5;", 5, Asterisk{}, 5},
+      infix_test_case<long>{"5 / 5;", 5, Slash{}, 5},
+      infix_test_case<long>{"5 > 5;", 5, GT{}, 5},
+      infix_test_case<long>{"5 < 5;", 5, LT{}, 5},
+      infix_test_case<long>{"5 == 5;", 5, Eq{}, 5},
+      infix_test_case<long>{"5 != 5;", 5, NotEq{}, 5},
       infix_test_case<bool>{"true == true", true, Eq{}, true},
       infix_test_case<bool>{"false == false", false, Eq{}, false},
       infix_test_case<bool>{"true != false", true, NotEq{}, false},
   };
   for (size_t i = 0; i < tests.size(); i++) {
     auto test{tests[i]};
-    auto int_test{std::any_cast<infix_test_case<int>>(&test)};
+    auto int_test{std::any_cast<infix_test_case<long>>(&test)};
     auto bool_test{std::any_cast<infix_test_case<bool>>(&test)};
     bool ret_val{true};
     if (int_test) {
